+++ test/mailers/previews/developer_mailer_preview.rb
@@ -1,0 +1,6 @@
+class DeveloperMailerPreview < ActionMailer::Preview
+  def invisiblize
+    notification = Notification.where(type: InvisiblizeDeveloperNotification.to_s).first
+    DeveloperMailer.with(record: notification, recipient: notification.recipient).invisiblize
+  end
+end